[Bronze V] Title: 이상한 기호, Time: 80 ms, Memory: 11620 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,8 +7,8 @@ class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         StringTokenizer st = new StringTokenizer(br.readLine());
         
-        int a= Integer.parseInt(st.nextToken());
-        int b= Integer.parseInt(st.nextToken());
+        long a= Integer.parseInt(st.nextToken());
+        long b= Integer.parseInt(st.nextToken());
         
         Main main= new Main();
         
@@ -17,7 +17,7 @@ class Main {
         bw.close();
     }
     
-    public int calc(int a, int b) {
+    public long calc(long a, long b) {
     	return (a+b)*(a-b);
     }
 }
